tests/wpt/web-platform-tests/web-animations/responsive/borderImageWidth.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id='container'>
<div id='element'></div>
</div>

<script>
var element = document.getElementById('element');
var container = document.getElementById('container');

test(function() {
    container.style.fontSize = '50px';
    var player = element.animate([{borderImageWidth: '300px 30px'}, {borderImageWidth: '10em 1em'}], 10);
    player.pause();
    player.currentTime = 5;
    container.style.fontSize = '10px';
    assert_equals(getComputedStyle(element).borderImageWidth, '200px 20px');
}, 'Border image width responsive to font size changes');

</script>
